채널 생성 다이얼로그 채널 이름 중복 체크 구현 - 중복된 이름일 경우 메시지 출력
