+++ app/asset/react_components/AppWrapper.jsx
@@ -3,6 +3,10 @@ import SideWrapper from './SideWrapper.jsx';
 import ContentWrapper from './ContentWrapper.jsx';
 import CreateChannelFormDialog from './content_components/CreateChannelFormDialog.jsx';
 import MessageWrapperComponent from './MessageWrapperComponent.jsx';
+import io from 'socket.io-client';
+import _ from 'lodash';
+
+const SOCKET = io('http://murmuring-ridge-75162.herokuapp.com/');
 
 class AppWrapper extends React.Component {
   constructor(props) {
@@ -11,6 +15,7 @@ class AppWrapper extends React.Component {
     this.state = {
       contentComponent: MessageWrapperComponent,
       openCreateChannelForm: false,
+      channelArr: [],
     };
     this.handleChnlAddClick = this.handleChnlAddClick.bind(this);
     this.handleChnlItemClick = this.handleChnlItemClick.bind(this);
@@ -29,6 +34,18 @@ class AppWrapper extends React.Component {
   handleCloseCreateChannelForm() {
     this.setState({openCreateChannelForm: false});
   }
+  componentWillMount() {
+    // 채널 리스트 받아오는 코드
+    this._getChnlInformation();
+  }
+  _getChnlInformation() {
+    SOCKET.on('roomInformation', (obj) => {
+      const chnlObj =  obj;
+      const result = _.keys(chnlObj).filter((key) => !(_.startsWith(_.trim(key), '/#')));
+      this.setState({channelArr: result});
+    });
+    SOCKET.emit('roomInformation');
+  }
   render() {
     return (
       <div>
@@ -39,12 +56,14 @@ class AppWrapper extends React.Component {
         <SideWrapper
           handleChnlAddClick={this.handleChnlAddClick}
           handleChnlItemClick={this.handleChnlItemClick}
+          channelArr={this.state.channelArr}
         />
         <CreateChannelFormDialog
           modal={false}
           open={this.state.openCreateChannelForm}
           onRequestClose={this.handleCloseCreateChannelForm}
           handleClose={this.handleCloseCreateChannelForm}
+          channelArr={this.state.channelArr}
         />
       </div>
     );
